+++ src/randomness.hpp
@@ -3,7 +3,7 @@
 
 #include <random>
 
-double RandFloat(double min, double max, std::default_random_engine &eng);
+double RandDouble(double min, double max, std::default_random_engine &eng);
 int RandInt(int min, int max, std::default_random_engine &eng);
 
 #endif // TERRAGEN_RANDOMNESS_HPP
